requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Caching
redis==5.0.1
fastapi-cache2==0.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

# URL validation and handling
validators==0.22.0
python-slugify==8.0.1

# CORS
fastapi-cors==0.0.6

# Environment variables
python-dotenv==1.0.0

# Date handling
pytz==2023.3.post1

# Performance monitoring
prometheus-fastapi-instrumentator==6.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Development tools
black==23.10.1
flake8==6.1.0
isort==5.12.0